requirements.txt
beautifulsoup4
PyGithub
requests